game/requirements.txt
tensorflow==2.20.0
keras==3.10.0
opencv-python
numpy<2
flask
openai
python-dotenv